requirements.txt
allure-pytest
allure-python-commons
enum34
execnet
namedlist
openapi-codec
packaging
pluggy
py
pyparsing
pytest
pytest-forked
pytest-html
pytest-metadata
pytest-ordering
pytest-xdist
requests
selenium
wcwidth
boto3
langdetect
webdriver-auto-update
pytz
pyautogui
urllib3
urllib3-secure-extra
webdriver-manager
testrail-api